Implement some import resolution rules
 - if the import path doesn't have an extension, see if it's short for
   a file with the extension '.js'
 - otherwise, treat it as a directory, and attempt to import index.js
   from that directory.

RxJs (or its esm2015 code) uses at least the first rule. The second
may be a NodeJS thing that I've lumped in, and I might get rid of it.

This is not quite adequate, because it will break on resolving nested,
relative imports, because as soon as a relative import is used as a
referrer, it will lose the base part of its path.