std: Provide an asArrayBuffer() method for flatbuffers
The communication with the go part is done through ArrayBuffers. However
flatbuffers will allocate a big array and build the flatbuffer starting from
the end of the ArrayBuffer and progressing backwards.

The result of that construct is a view of the ArrayBuffer, not the full length
of it. When passing the underlying array buffer to go, it's enable to find the
start of data.

We need to splice (copy!!) the flatbuffer to a smaller array buffer for go to
make any sense of it.