Add basic machinery for deferreds
To have functions in JavaScript that will call into Go code to do I/O,
we need a bit of bookkeeping and a bit of protocol.

This commit adds some of the required machinery, in the package
`deferred`. This is wired into the runtime, and used in the
implementation of a `read` built-in.

Lacking:
 - the JavaScript side of the machinery
 - ability to cancel an outstanding request
 - a non-silly implementation of read
 - non-silly scheduling of requests