Use [ubyte] for data, rather than string
In flatbuffers, `string` means a UTF8 string, which is not
interchangeable with an array of bytes. For file contents, it's more
appropriate, and will save heartache when dealing with checksums,
etc., to send bytes as bytes. In flatbuffers that's a `[ubyte]`
(vector of unsigned bytes), and happily it does generate some
convenience methods for setting and getting useful values in both Go
and JavaScript.

This means we have to do a tiny amount more work if we want to write a
string back out. `write` ought to accept a byte array; but for now, it
expects a string, and it's trickier to account for JSON-encoded things
if we have to send them as bytes. Though not impossible.