Include name and path in filesystem return values
It's convenient to have the path to a file or directory given to you
rather than having to construct it, since the filesystem library
functions _accept_ paths. But occasionally (e.g., for constructing
ConfigMaps, where the keys are the filenames) it's handy to have the
name right there too. So, return both.