Add Encoding argument to read()
If we want to be able to return different formats from read (e.g.,
readJSON), then the fundamental thing is to be able to request and
receive unambiguously-encoded strings. UTF8 (which is natural to
provide from Go) is no good, because it'd be a pain to decode in
JavaScript. UTF16, on the other hand, it knows how to do.

This commit adds support for Encoding in the read RPC, so that the Go
part of the runtime can encode data as UTF16, and the JS side can
request that, and decode it as such.