Factor out Format enum for use in read as well
As well as specifying the format to write in, we may want to specify
the format to expect when reading. In order to use the same enum, it
needs to be included, and the most sensible way to arrange this is to
put it in its own file.

Big however: when the flatbuffers JS generator is invoked on each
file, it will generate a distinct module each time. The generated code
assumes all definitions are available, but this is not the case --
they may be in another module.

The solution is to generate a single JavaScript file, and therefore a
single module, for the __std package.

Lastly, a relatively minor change: call the Format-typed WriteArgs
argument `format`, for clarity.