Add ability to read structured data as JS value
It'll often be useful to read a file in and parse it into an
object. Rather than require e.g., YAML parsing on the JavaScript side,
give people the option of asking for `JSON` encoding, which means
"parse whatever you read, and send a JSON object as a string".

Not all combinations of encoding and format make sense:

 - the format argument is only used if the encoding is `JSON`; the
   other encodings don't interpret the contents, so there's no reason
   for them to care about the format.

 - asking for encoding.JSON, format.Raw doesn't have a sensible
   meaning, so it's treated the same as Format.JSON

An alternate design for this would be to simply include the expected
formats as encodings, i.e., so they are Bytes, UTF16, YAML, JSON,
Auto.