Separate deferred.Register and RegisterWithContext
Give two ways to register a deferred value: one in which the context
is supplied, and the operation can be cancelled; and one in which the
operation just uses the background context.

The intention is that being cancelable is "user pays". But it doesn't
make much sense for it to be a user-supplied value, as I've done here.
probably just keep the mechanism, for future use, and not expose it in
read().