Default read() to returning a JS value
A slight tweak, to make the `encoding` argument less awkward: treat it
as the desired form of the return value: either the bytes, a string,
or a (JSON) value. The `format` is used to override the expected form
of the read contents, if necessary. (Raw still doesn't make much
sense, so it's conflated with JSON).