Use resolved path when resolving relative imports
Prior to npm module resolution, the specifier for a module and the
file path had a direct relationship, and it was a reasonable
approximation to make base/specifier the new base for recursive
imports.

However, npm modules can be resolved to paths that don't correspond
closely with the specifier; so, explicitly return the filesystem path
from `Import` and use that to resolve recursive imports.