Supply a fallback promise rejection handler
V8Worker2 will look for a global called `onerror`; so load one before
starting.

Also: without intervention, errors that come from the "deferred"
system (when the runtime does something for you and returns the result
later) will have a helpful error message (because we pass it back to
JS), but a meaningless stack trace (because it starts where control
re-enters JS, i.e., at `recv`). To make it a bit more useful, capture
the stack trace when entering code that asks for a deferred value, and
transplant it to any eventual error.