std: Make log() a bit more clever when dealing with strings
The current behaviour of printing JSON values isn't ideal when printing a
string. We have:

std.log('foo') => "foo"

We special case strings so we don't end up with the double quotes. It's really
confusing and hard to work around.

A last detail: when asked to write JSON though, we make sure to still print
strings as JSON values!