Introduce exec option for root read directory
This adds an option (not a flag, yet) for the runtime to tell it where
the root directory for reads. The idea is that relative paths will be
interpreted with respect to the root path.

This opens the way to providing a directory _other_ than the current
working directory -- e.g., that of the script being run.

If we are being strictly hermetic, absolute paths should be
disallowed. For the minute, I've just restricted them to paths that
happen to be under the root path.