Give `jk run` flags to set input directory
This adds two flags to the `jk run` command:

 * --directory,-C means use the given directory as the base for reads

 * --use-script-dir means use the directory of the script argument as
     the base for reads

The default is to use the current working directory as the base.

If you're running a library script against your set of files (and they
are not in $PWD), `--directory` is what you want; if you're running a
script that lives with its own files (and they are not in $PWD),
`--use-script-dir` is what you want.