Correct off-by-one(-dir) mistake in NPM importer
The NPM importer assumed it would be passed a script name as the base
path, and always dropped the last element by using `filepath.Dir` on
the base. This breaks when it's not the trivial case of the base being
the working directory, since the resolver already drops the last
element (it works in the trivial case because `Dir('.')` is `'.'`).