Change input dir flag and provide tests
Instead of the default input directory being the current working
directory, make the default the directory containing the script to be
run. That way, the default is still a sensible one, and there's only
one flag needed.

It's not quite appropriate to use the Unix tradition of -C denoting a
working directory, so use `--input-directory,-i`. If there's a
sensible interpretation of -C, it can be done later.