param: Fix NPE when not specifying a default Object value
We had a bug for Object parameters specifically. For objects, we deep merge the
object default value with param sub-subtree that the user already provided and
so we need to send the object default value to the go side.

We were serializing the object default value as a JSON string for that but
without checking the user did specify a default object.

Fixes: #122