Sketch out resource module
This implements a "magic" module '@jkcfg/std/resource'. It's magic
because it's generated each time it's imported, so that it can take
into account the path of the importing module (known at import time,
and not thereafter).

The implementations is this: create a module that embeds the base path
of the module, and exports a procedure that does reads relative to the
base path.

This has some flaws:

 - at present, jk (v8worker2, really) gives imports using the same
   specifier (in this case `@jkcfg/std/resource`) the same module,
   which is not much use for magic modules.

 - it relies on the module base path being readable (if you change the
   input path when you run jk, it might not be)