Test that std/resource not subject to input dir
This commit adds a test that resource() can succeed even if the module
directory is not under the input directory. The test does not pass,
since all reads are currently required to be under the input
directory.

It also corrects the guard against reading outside of the input
directory: it's possible for os/path/filepath.Rel to return a relative
path that starts with parent directories, i.e., with `..`.