Ensure resources are relative to the module path
Just using std.read() for resources won't work in general, because the
desired path may not be under the input directory (and reads are
restricted to paths under the read directory).

To enable reads from module directories, while keeping the guard
against reading outside the allowed paths, keep track of each module
that imports `@jkcfg/std/resource' and look up the module path when
its `resource` procedure is called, to use as a base path.