Add salt to module hashes
To be able to identify the base path of a resource read, I added a
`module` option to `read`. That is now "generally available", meaning
any read might include it, if it can guess a module hash, and read
from some other module's base path.

To prevent abuse of this mechanism, salt the module hash using random
bytes generated when starting up. The randomness does not hinder
repeatability, since the hashes are opaque and only used in the scope
of a run anyway.