Add merge function and unit tests for merge module
This adds a `merge` procedure, with semantics similar to object merge
in jsonnet, to the ported std_merge library.

Suddenly we have a module that does not depend on the runtime; so this
commit also sets things up so jest runs unit tests as part of `make
test` (or `npm test` in std/). eslint on std/ is now done as its own
step in the run-test.sh script.

Regrettably, using jest (or babel?) pulls in an internet of npm
packages.