std: Serialize resolution of deferred functions
We want the js author to be able to rely on deterministic order of promise
resolution to avoid surprises and generate non-determinitic configuration.

We do this by making the order of promise resolution be the same as promise
resolution, serializing the deferred goroutines before they re-enter v8 to
resolve the promises.

Fixes: #110