Try import resolution to files before directories
The npm resolution rules try to resolve a path as a file before
resolving as a directory. I had taken a shortcut by just checking
whether the path existed and was a file, or a directory. But this
neglects to check the path+extension rules, such that

    foo/

was checked in preference to

    foo.js

in the case that they both exist. This led to getting the wrong module
(if say, `foo/index.js` exists) or wrongly failing resolution (if
`foo/` and `foo.js` exist but not `foo/index.js`).