std: Use modules in the standard library
Instead of using rollup to transform the std lib code into a single module we
now bundle the stdlib modules as they are written in jk.

For the import to work, StdImporter is a special importer that understand the
std modules are built-in assets. There are a few interesting details:

- We need to resolve modules relative to std lib modules
- We white list the modules the final user can import from the std lib. Other
  modules are considered private and can only be importer from within the std
  lib itself.