std: Generate a @jkcfg/std modules with typings
A big part of the goal of transforming the std lib to typescript is to be able
to generate typings directly from the source code instead of having to
duplicate the function signatures in a shim module.

This commit does just that: remove std/shim in favour a full module generated
from the std/. make module will create a @jkcfg/std directory, ready to be
published.