run: Add an option to emit the list of a script dependencies
It can be useful for external programs to know the list of files a script
depends on. For instance, a program can watch that list of files and run jk
when one of them changes as it's possible running the script will produce a
different output.

We current emit only file dependencies:

  - import-file: a file imported through an import statement
  - read-file: a file read by std lthe output has changed as well.ib

Fixes: #172