Let merge be its own module
Previous changes made it possible to import e.g., `@jkcfg/std/fs` when
running `jk`; but the NPM package includes things as e.g.,
`@jkcfg/std/std_merge`. This commit renames the modules in std/ so
that they will be importable from the NPM package the same as from the
runtime.