Ensure go.{mod,sum} are not modified during builds
Usually, `go build` etc., will alter go.mod and go.sum as needed;
e.g., by adding modules that are imported but not yet mentioned there.

This can mean that a CI build adds modules that aren't pinned down --
and that the record of what was used will be lost when the build
fileysystem is torn down.

To make sure that changes are deliberate, I've added the flag
`-mod=readonly` to go invocations (including in code generation). This
will make them fail if go.mod or go.sum would be modified -- i.e., it
makes those files behave more like a lock file.

To update or add dependencies, invoke go by hand, e.g., as `go mod
tidy`.

To make go invocations properly deal with the tools needed by the
generate script, I've imported those from a file, but given it a build
tag of `tools` so its not included in the main build.