Add generic RPC mechanism to protocol
To support arbitrary new operations without needing to extend the
protocol between the runtime and JavaScript, this commit adds an RPC
message to that protocol.

In JavaScript the API is `std/internal/rpc#RPC(method, ...args)`,
which will return a Promise. In the runtime, the message is handled in
std.go, at present with a stub that just prints the arguments sent.

Arguments are encoded in two different ways: most value will be
serialised as JSON, but Uint8Array arguments (i.e., byte buffers) are
sent as a byte buffer, since the JSON serialisation of such would be
terrible. (It's possible this doesn't need all the union/table faff in
flatbuffers, since strings and `[uint8]` in flatbuffers both look like
`[]byte` in Go, so could be punned, with just a discriminator field).