Provide for sync RPC as well
This adds a flag into the protocol to call an RPC method
synchronously, and getting the result directly rather than a
promise. At present, this is left to the JavaScript to decide (so you
can call any method in the dispatch table either synchronously or
asynchronously); in the future we may want to distinguish between them
in the runtime.

I also refactored the flatbuffer schema minorly, so I could use the
Data type again in synchronous RPC responses.