Port std/fs to use RPCSync
We can now use the RPC mechanism as the means of calling std
functions, thereby removing a lot of specialised serialisation code.

This commit reimplements fs/info and fs/dir as RPCs. RPC dispatch is
now done on two levels: the std library has its own hard-wired set of
methods, and falls back to a table of extensions provided by the
runtime.