Handle multidocs in jk transform
This makes `jk transform` transparently apply its transformation to
all the objects in a YAML or JSON multidoc.

I realised that once the inputs are transformed, we are essentially in
the same situation as `jk generate` -- we have a bunch of objects that
need to be output, and hints as to their format. For that reason, I've
factored that out of the JavaScript side of generate, so it can be
called from transform.