Provide the option to error on overwrite
Previously the choice for a write that would overwrite a file was to
skip it, or overwrite the file. For `jk transform` (and possibly other
things), the ideal is to throw an error unless it's been explicitly
allowed.

This commit makes the overwrite field an enum, and changes the runtime
protocol to return an error if that overwrite option is used. A
boolean is also accepted in the JS API, for backward compatibility.