Default jk transform to file output
Rather than having --stdout=true be the default, meaning you need to
supply --stdout=false if you want to write to files, it is more
consistent with `jk generate` to assume file output.

To keep it safe and clean, the default is _also_ to raise an error
rather than overwriting.