Enable large string values from runtime
When reading a file, the value is (usually) sent back to JavaScript
encoded as UTF16, to make it easier to turn into a string --
JavaScript uses UTF16, and `String.fromCharCode` expects 16 bit values
(in unicode taxonomy: code units).

`String.fromCharCode` is tremendously convenient for this purpose,
_except_ that it accepts varargs (rather than an array of code units),
and there's a practical limit to the number of varargs that can be
supplied. Reading file larger than a few tens of kilobytes will exceed
that limit.

So: do the conversion of bytes to strings in chunks, sticking to a
conservative number of varargs at a time. This also changes from
`String.fromCodePoint` to `String.fromCharCode` (which expects code
_units_, i.e., the 16-bit values, which is what we have).

I have not made `stringFromUTF8Bytes` work the same way; it is used
for decoding values _constructed_ by the runtime; e.g., the results of
RPC calls. These are less likely to be large in size -- though of
course the possibility exists, and it could be handled the same way.