Implement procedures as in RFC
This implements the three procedures `validateBySchema`,
`validateByFile`, and `validateByResource`, as described in the JSON
Schema validation RFC, with some differences:

 - `validateBySchema` is left synchronous since it isn't expected to
   involve arbitrary I/O (but could be made asynchronous to match the
   others).

 - `validateByResource` requires a "module ref" (i.e., must be called
   via `@jkcfg/std/resource#withModuleRef`); more on this below.

The latter difference is to fit with how resources work. The magic
`@jkcfg/std.resource` import grants access to module-relative files by
generating code that refers to the module path, and exporting
functions that close over it. This commit adds a function
`withModuleRef` so that a module can supply the module ref, and
thereby grant access to its resources, to `validateByResource`.