Provide filesystem walking functions
It is really useful to be able to recursively walk the filesystem, and
a natural way to express this in JavaScript is with generators.

There are many formulations of `walk`; for now I have hedged my bets
and given two -- one that has a directory-oriented view (a bit like
Python's os.walk), and one that just gives you each file info.