Walk in preorder traversal
It's easy to just have one walk procedure and make it yield file
infos, which after all have the information you most need (path, name,
and whether it's a directory). But it'd be nice to be able to know
which files are in which directory, without having to do fiddly
calculations with the path.

Doing a preorder traversal at least means you will see a directory
immediately before you see its contained files. You still cannot tell
where you are in the tree (after all, that's what a traversal is --
linearising).