Enrich errors with structure
This commit makes the validation errors returned from the runtime into
structs, rather than simple strings, so that the JavaScript caller can
decide how to format them.

Although the location is included in the struct, it is not returned by
the schema validation library, so it's not included in the
accompanying Go code.

This also adjusts the naming of the JavaScript schema validation
functions, as suggested in the PR comments.