Provide print and change log
This changes `std.log` so that it uses `V8Worker2.log`, and thereby
outputs to stderr.

Before log was subordinated to the role of logging, it was used
extensively to print simple results to stdout, for testing and
whatnot.

The need for printing to stdout is common; so rather than rewrite
everything to use

    write(val, '', opts)

this also adds a `print` procedure which can be used instead.