Implement module search path
This adds a flag `-I` to all the script-executing subcommands of
`jk`. The flag adds a directory to the module search path; that is,
the directories from which imports are resolved.

A natural implication is that if I use `-I ~/lib/jk`, then a module at
`~/lib/jk/foo/index.js` can now be imported as `'import f from
'foo'`. Making this the case needed a bit of finessing, since the
FileImporter treated everything as relative to the path of the
importing module. Now, _only_ paths beginning with '../' or './' are
treated as relative to the importing module (which is how the NPM
resolution works already).