Use virtual filesystems for access to filesystem
The essential change here is that almost everything that used a path
in the host filesystem now uses a path _within_ a virtual
filesystem. The purpose is two-fold:

 - it restricts file access to the top of the filesytem, so there's no
   need for trying to figure out whether they "escape" their sandbox

 - now we can read from things that aren't just directories in the
   host filesystem

This requires a change from `filepath` to `path`, since paths in
http.FileSystem are always forward slashed (except for filepath.Rel,
for which there is no path counterpart).

To properly account for relative specifiers (e.g., `./foo`), there
must be an importer that specifically looks in the importing module's
filesystem. This is a cleaner arrangement than before, when each
importer checked whether there could be a path relative to the
importing path.