Distinguish between internal and user filesystems
It is no longer possible to decide whether an import should be
recorded based on the kind of importer, since Relative will examine
whichever filesystem is doing the importing, and that might be an
"internal" source.

However, each _filesystem_ is either supplied by the user (implicitly)
or integral to the system, so that distinction can be used instead.