Detect forbidden paths with path.Clean
Using an http.FileSystem as the base directory for reads means that
paths outside it will simply not be found. But we do want to provide a
more specific message in the case of someone accidentally trying to
read outside the input directory or module, so it's still useful to
try and detect that.

It is no longer convenient to calculate using `filepath.Rel` since we
don't have the absolute path to the base (and it might not be a
filesystem directory anyway). That was a bit suspect, in any case,
since it is specific to the OS (i.e., uses backslashes and volume
names on Windows), whereas `jk` expects forward-slashed paths whatever
the venue.

Instead, use this mostly-accurate rule: a forbidden path will start
with `..` when cleaned. The mechanical cleaning will put any parent
paths at the start of the string. As noted, it's possible to have a
path that begins with `..` but happens _not_ to be forbidden, but this
can always be rewritten.