Distinguish Make{FileInfo,DirectoryListing}
Rather than have these as methods of ReadBase, accept ReadBase as an
argument. This opens the way to ReadBase being an interface rather
than a concrete type, so different implementations might be used.

Also: rename each to Make<X>, so that they have distinct names to
their return value types.