Handle multidocs in transform
Per #332, transform didn't treat multidocs correctly -- while it read
them in OK, it made a separate entry for each value in the files to
write out, which would trigger the overwrite protection.

The correct thing to do is to transform each value, but keep the
stream as an array of values, and forward that along with the stream
format to the generate code, so it'll write them out as a stream to a
single file.

However, doing that exposes another problem: when printing results to
stdout, generate wants everything to be the same format (and either
JSON or YAML). It should really handle stream values by inlining them,
assuming it's _compatible_ with the other values.

Fixing that latter needed some more intrusive changes, to make
generate assemble what it's going to write out and check the format as
it goes.