Use a magic module to expose the host filesystem
To be able to process the files named on the command line, for `jk
validate` and `jk transform`, and no doubt other command to come, the
standard library needs access to the full host filesystem.

This commit makes the whole filesystem readable by using the same
"magic module" mechanism as for resources, but exposing it only to
modules from the standard library (as `@jkcfg/std/internal/host`).

It still won't be possible to read from absolute paths or parent
paths, so that needs adjusting next.