Use the host module from validate and transform
The modules that implement `jk validate` and `jk transform` can now
get access to the filesystem with the magic import
`@jkcfg/std/internal/host`. This commit makes that work, by:

 - providing TypeScript definitions for the magic module, and making
   sure tsc can find them (this makes it able to find any absolute
   import from '@jkcfg/std');

 - making sure the input directory for the VM is an absolute path, so
   that file paths in the host (real) filesystem are resolved
   correctly.

Previously, this latter worked by coincidence because "." (the working
directory if not set) was also the root of the virtual filesystem. Any
path like 'foo/bar' would be made into '/foo/bar' and read in the VFS
rooted at '.', and so get the right file.

But with the host filesystem, the root is `/` and the base path is the
working directory. Adding 'foo/bar' to path '.' tries to read the file
/foo/bar in the host filesystem.