Allow absolute paths for some locations
To let validate and transform deal with files anywhere in the
filesystem, I need reads to be able to allow absolute (and parent)
paths -- _sometimes_.

All reads (and filesystem things in general) go through
ReadBase.getpath; but the absolute path handling depends on the
particular base location, and may vary per module. The most
straight-forward thing to do is mark a particular vfs.Location as
allowing absolute paths, and account for that in ReadBase.

I cleaned up getPath by returning the path within the vfs as a
Location, rather than as a separate return value. Most callers were
doing path.Join themselves; this saves that.