Use const symbols for stdin and stdout
Up to now, passing `''` has meant stdin (for read) and stdout (for
write), but this is not very guessable.

Instead, use a symbol for each of `stdin` and `stdout`. The
combo-keyword `unique symbol` in TypeScript means that `typeof stdin`
will accept only that symbol (in contrast with the type `symbol`).