Chroot filesystem in images
Instead of expecting modules to be dumped in the root directory, it's
friendlier to give jk a "namespace" by putting things in a
subdirectory. For modules, I have chosen `/jk/modules`.

But to make sure we look for files in the right location (and not
outside it!), we have to do the same thing as we do for modules in the
"real" filesystem, and treat the chosen directory as the root of the
image filesystem. This requires an extra layer, to "chroot" in the
image filesystem.