Allow parent and absolute paths for transform
When you call `jk transform --overwrite`, the expectation is that jk
will overwrite the files you mention. But if these have absolute or
parent paths, the sandbox will forbid the write, so there must be a
mechanism for allowing writes to the exact files.

 - add a test that transform can write to a parent path that's
   mentioned as an input, when overwriting

`jk transform` and `jk validate` need to read files that are mentioned
explicitly on the command-line, so got a sandbox escape mechanism in
PR #338, in which a location can be marked as allowed to escape the
sandbox.

A better place to put it is with the record that's registered for each
module. Then, a module can also be marked as allowing writes to the
host, so I can:

 - add a `module` argument to write;
 - give the magic std/internal/host module a write function;
 - adjust the sandbox so that modules can be marked as allowed to
   write to the host; and,
 - use host.write to write files in std/cmd/transform

.. which, finally, lets transform overwrite the files it's given on
the command line.