Test sandboxing
There are two flags recorded for a module:

 - AllowPathsOutsideSandbox (whether you can use parent or absolute paths)
 - AllowWritesToHost (whether you can write with this module).

All writes are to the host filesystem, so the latter really means "can
you do writes using this module", and in combination with the former,
"can you do writes outside the output directory".

The tests cover the four different operations

    {read,write} x {inside,outside} the sandbox

and the four combinations of flags, as well as the case of _not_
reading and writing via a module.