Let the user force the output format for generate
This adds a flag `jk generate --format`, which will force the output
to either YAML or JSON.

If printing to stdout, this means all values are printed as whichever
foramt, with streams being inlined as before.

If writing to files, the file extension when present will be
changed. Values with a stream (or multidoc) format -- YAMLStream,
JSONStream -- will still be written as streams.